moving aotf playground to another component
