+++ apps/platform/src/pages/DiseasePage/DiseaseAssociations/DiseaseAssociations.jsx
@@ -14,34 +14,10 @@ import {
   DataUploader,
 } from "../../../components/AssociationsToolkit";
 import DISEASE_ASSOCIATIONS_QUERY from "./DiseaseAssociationsQuery.gql";
-import { ApiPlaygroundDrawer } from "ui";
+import AotfApiPlayground from "../../../components/AssociationsToolkit/AotfApiPlayground";
 
 function AssociationsWrapper() {
-  const {
-    initialLoading,
-    id,
-    pagination,
-    searhFilter,
-    sorting,
-    enableIndirect,
-    dataSourcesWeights,
-    entity,
-    dataSourcesRequired,
-  } = useAotfContext();
-
-  const aggregationFilters = dataSourcesRequired.map(({ id, ...obj }) => ({ ...obj }));
-
-  const variables = {
-    id,
-    index: pagination.pageIndex,
-    size: pagination.pageSize,
-    filter: searhFilter,
-    sortBy: sorting[0].id,
-    enableIndirect,
-    datasources: dataSourcesWeights,
-    entity,
-    aggregationFilters,
-  };
+  const { initialLoading, id } = useAotfContext();
 
   if (initialLoading) return <AotFLoader />;
 
@@ -57,10 +33,7 @@ function AssociationsWrapper() {
             </PrivateWrapper>
             <Divider orientation="vertical" />
             <DataDownloader fileStem={`OT-${id}-associated-targets`} />
-            <ApiPlaygroundDrawer
-              query={DISEASE_ASSOCIATIONS_QUERY.loc.source.body}
-              variables={variables}
-            />
+            <AotfApiPlayground />
           </OptionsControlls>
         </Box>
         <Box>
